Separacion de Proxys en distintos archivos
